:construction: Fix left/right movement for the most part
There is currently a problem with ratios that would allow for irrational
fractions.  So, if you have, for example, ratios of 1 1 1, then despite
all of the sizes technically being the same, they actually fall on the
LAST widget, rather than the first, because of rounding.

I'll take a look at that and if it is all good then I can work on the
other parts of modularity.