feature: Show process state
This is not 100% finished and will be refined in the future, as I plan to
do a bit of an overhaul on how the process widget is going to look and
functionality.  In particular, tabbed is currently kinda just slapped
together (I just combine all the states together as one big string).

However, it is enough to work and show state normally...