refactor: Cut out sysinfo from Linux
Refactors to use only heim for Linux builds.  This is now much easier to
do since the 0.1 version of heim works fine for ARM.  This is ideal
since having to rely on two separate sources of data isn't the greatest
if we can avoid it.

Sysinfo is still required for macOS and Windows, though.  Temperature
sensors do not work for those from heim, and for some reason, networks
also don't work on Windows with heim...?

My personal CPU core calculation is also currently Linux-only, and as
such, I'll still rely on sysinfo for Windows and macOS for now.

This isn't really a big optimization or anything btw.