feature: Hide SWAP graph and legend in normal mode if SWAP is 0
Firstly, note this currently won't affect basic mode.  There is
code changes due to it, but instead, we'll just display `0B/0B`
instead.  I'm personally not really sure if we want to get rid
of it in basic mode, since it'll leave an ugly gap in that mode.

Anyways, this change is mainly for the normal mode.  All this does
is hide the legend entry and chart if the total SWAP drops to 0 KB.
That's it.