bug: fix endless loop when resizing the window to a width that is too small
Due to a missing check, you could resize the window to a width that was
too small, and it would trigger an endless while-loop for any table
while trying to redistribute remaining space. This has been
rectified with an explicit check, as well as a smarter method of
redistributing remaining space borrowed from the rewrite.

Note that minimal heights don't seem to be any issues.