Added custom terminal title
I added a new option in terminal and in the config
file to be able to change the terminal title with
any custom one. The user can now add `--title` and
the custom title after it or add the `title_to_hostname`
field in the config file to set the terminal's name
to hostname. If there is no option found, then
the name of the terminal will be set to "Bottom".